MLIBZ-2417 Move IOSClientBuider to separate class. Implementation of device info header for Android.
