+++ Kinvey-Xamarin-Android/Kinvey-Xamarin-Android.csproj
@@ -57,6 +57,12 @@
     <Reference Include="Newtonsoft.Json">
       <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\portable-net45+wp80+win8+wpa81\Newtonsoft.Json.dll</HintPath>
     </Reference>
+    <Reference Include="SQLite.Net">
+      <HintPath>..\packages\SQLite.Net.Core-PCL.3.1.1\lib\portable-win8+net45+wp8+wpa81+MonoAndroid1+MonoTouch1\SQLite.Net.dll</HintPath>
+    </Reference>
+    <Reference Include="SQLite.Net.Platform.XamarinAndroid">
+      <HintPath>..\packages\SQLite.Net-PCL.3.1.1\lib\MonoAndroid\SQLite.Net.Platform.XamarinAndroid.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Push\KinveyGCMService.cs" />
@@ -65,6 +71,7 @@
     <Compile Include="Auth\AndroidNativeCredentialStore.cs" />
     <Compile Include="Auth\KinveyAccountService.cs" />
     <Compile Include="Auth\UserExtension.cs" />
+    <Compile Include="Client\AndroidClientBuilder.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -74,6 +81,7 @@
   <ItemGroup>
     <Folder Include="Push\" />
     <Folder Include="Resources\" />
+    <Folder Include="Client\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Kinvey-Utils\Kinvey-Utils.csproj">
