+++ Kinvey-Xamarin-Android/packages.config
@@ -2,6 +2,8 @@
 <packages>
   <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="MonoAndroid50" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="monoandroid70" />
+  <package id="SQLite.Net.Core-PCL" version="3.1.1" targetFramework="monoandroid70" />
+  <package id="SQLite.Net-PCL" version="3.1.1" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.v4" version="23.1.1.0" targetFramework="MonoAndroid50" />
   <package id="Xamarin.GooglePlayServices.Base" version="27.0.0.0" targetFramework="MonoAndroid50" />
   <package id="Xamarin.GooglePlayServices.Basement" version="27.0.0.0" targetFramework="MonoAndroid50" />
